+++ src/Commands/Env/DeployCommand.php
@@ -3,6 +3,7 @@
 namespace Pantheon\Terminus\Commands\Env;
 
 use Pantheon\Terminus\Commands\TerminusCommand;
+use Pantheon\Terminus\Exceptions\TerminusException;
 use Pantheon\Terminus\Site\SiteAwareInterface;
 use Pantheon\Terminus\Site\SiteAwareTrait;
 
@@ -31,6 +32,8 @@ class DeployCommand extends TerminusCommand implements SiteAwareInterface
      * @option string $updatedb Run update.php after deploy (Drupal only)
      * @option string $note Custom deploy log message
      *
+     * @throws TerminusException
+     *
      * @usage <site>.test Deploy code from <site>'s Dev environment to the Test environment.
      * @usage <site>.live Deploy code from <site>'s Test environment to the Live environment.
      * @usage <site>.test --cc Deploy code from <site>'s Dev environment to the Test environment and clear caches on the Test environment.
@@ -42,7 +45,7 @@ class DeployCommand extends TerminusCommand implements SiteAwareInterface
         $site_env,
         $options = ['sync-content' => false, 'note' => 'Deploy from Terminus', 'cc' => false, 'updatedb' => false,]
     ) {
-        list(, $env) = $this->getUnfrozenSiteEnv($site_env, 'dev');
+        list($site, $env) = $this->getUnfrozenSiteEnv($site_env, 'dev');
 
         if ($env->isInitialized()) {
             if (!$env->hasDeployableCode()) {
@@ -56,8 +59,14 @@ class DeployCommand extends TerminusCommand implements SiteAwareInterface
               'annotation'  => $options['note'],
             ];
             if ($env->id == 'test' && isset($options['sync-content']) && $options['sync-content']) {
-                $params['clone_database'] = ['from_environment' => 'live',];
-                $params['clone_files']    = ['from_environment' => 'live',];
+                $live_env = 'live';
+                if (!$site->getEnvironments()->get($live_env)->isInitialized()) {
+                    throw new TerminusException(
+                        "{site}'s {env} environment cannot be cloned because it has not been initialized.",
+                        ['site' => $site->getName(), 'env' => $live_env,]
+                    );
+                }
+                $params['clone_files'] = $params['clone_database'] = ['from_environment' => $live_env,];
             }
             $workflow = $env->deploy($params);
         } else {