+++ CHANGELOG.md
@@ -7,9 +7,11 @@ All notable changes to this project will be documented in this file. This projec
 
 ### Changed
 - `self:cc` now acts to delete all files in the command cache directory. (#1569)
+- `env:clone-content` and `env:deploy` now refuse to clone from uninitialized environments. (#1608)
 
 ### Deprecated
 - The `element` parameter on `backup:list` is deprecated. Use the `--element=` option instead. (#1563)
+- The `wait()` function of `Workflows` is deprecated. Use `checkStatus()` instead. (#1584)
 
 ### Fixed
 - Fixed the base branch in the URL when checking for upstream updates. (#1581)